+++ duilib/Control/ListCtrlReportView.cpp
@@ -1518,7 +1518,7 @@ void ListCtrlReportLayout::LazyArrangeChild(UiRect rc) const
 
     //计算当前区域能够显示多少条数据
     int32_t nCalcItemCount = pDataView->GetMaxDataItemsToShow(nScrollPosY, rc.Height());
-    if (nCalcItemCount > (nItemCount - 1)) {
+    if (nCalcItemCount > (int32_t)(nItemCount - 1)) {
         //UI控件的个数不足，重新调整
         pDataView->AjustItemCount();
     }
